requirements.lock.txt
annotated-types==0.5.0
anyio==3.7.1
atproto==0.0.46
certifi==2025.11.12
cffi==1.15.1
charset-normalizer==3.4.4
click==8.1.8
configparser==5.3.0
cryptography==41.0.7
cycler==0.11.0
dnspython==2.3.0
exceptiongroup==1.3.1
fonttools==4.38.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
importlib-metadata==6.7.0
kiwisolver==1.4.5
libipld==1.2.3
matplotlib==3.5.3
numpy==1.21.6
packaging==24.0
Pillow==9.5.0
pkg_resources==0.0.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyOpenSSL==23.2.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
requests==2.31.0
RPi-AS3935==0.2.0
RPi.GPIO==0.7.1
six==1.17.0
smbus2==0.5.0
sniffio==1.3.1
typing_extensions==4.7.1
urllib3==1.26.18
websockets==11.0.3
zipp==3.15.0
